refactor(db): renumber migrations + extend products/carts schema
Renumber the migration files to a clean 0001..0009 sequence so the
runner's lexical order matches the dependency order. The previous
schema-bake-and-amend cycle had left gaps (0003, 0010, 0013, 0015, 0022,
0029, 0030, 0041, 0043, 0098) — readable diffs are easier with one
contiguous range.

Schema extensions for the storefront:
- products: add storefront-facing fields (image_url, weight_grams, …)
  baked into 0000000004-create-products-table.sql instead of carrying a
  follow-up alter migration.
- carts: gain checkout state columns (email, shipping_*, checkout_step,
  session_token, status) so the multi-step checkout actions have
  somewhere to write the running form state. Baked into
  0000000008-create-carts-table.sql.

Cart and Product models grow to match (79 + 31 lines of attributes /
factory data). ProductsSeeder + products.sql refreshed for the dog-
treat catalog used in the demo storefront. .qb model snapshot updated.

+++ database/migrations/0000000002-create-order_items-table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS "order_items" (
   "id" INTEGER PRIMARY KEY AUTOINCREMENT,
-  "quantity" REAL default 1,
-  "price" REAL,
+  "quantity" INTEGER default 1,
+  "price" INTEGER,
   "special_instructions" TEXT,
   "order_id" INTEGER,
   "product_id" INTEGER,
